refactor(workflow): separate database and agent activities
