+++ src/workflow/activities/index.ts
@@ -2,10 +2,12 @@ import * as taskActivities from './task'
 import * as aiActivities from './ai'
 import * as transcodeActivities from '@/transcode/activities/transcode'
 import * as agentActivities from './agent'
+import * as dbActivities from './db'
 
 export const activities = {
   ...taskActivities,
   ...aiActivities,
   ...transcodeActivities,
   ...agentActivities,
+  ...dbActivities,
 }
